🛠️ Refactor: Extract domains, eliminate magic numbers and fix silent errors
- Extracted monolithic `revancedbot/__init__.py` into domain-specific modules (`models.py`, `app.py`, `fetcher.py`, `patcher.py`, `cli.py`, `errors.py`), justifying Single Responsibility Principle and Extract Class patterns.
- Replaced magic numbers with named constants (`WINDOW_WIDTH`, `WINDOW_HEIGHT`, `POLL_INTERVAL_SECONDS`, `SETTLE_DELAY_SECONDS`, `LOW_LIMIT_MAX_JOBS`).
- Fixed silent error swallowing in `app.py` by implementing a centralized `report_error` function in `errors.py`.
- Formalized architecture locations and error reporting guidelines in `AGENTS.md`.
