+++ revancedbot/models.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+from typing import Optional
+
+@dataclass
+class PatchJob:
+    package_id: str
+    package_version: Optional[str] # latest if None
